Fix rolegreets, add other neat stuff

+++ src/Mewdeko.Database/Models/RoleGreet.cs
@@ -7,7 +7,7 @@ public class RoleGreet : DbEntity
     public ulong ChannelId { get; set; }
     public bool GreetBots { get; set; }
     public string Message { get; set; } = "Welcome %user%";
-    public ulong DeleteTime { get; set; } = 0;
+    public int DeleteTime { get; set; } = 0;
     public string WebhookUrl { get; set; }
     public bool Disabled { get; set; }
 }
